requirements.txt
# ══════════════════════════════════════════════════════════════
# 위밋모빌리티 해상 리스크 플랫폼 — Python 의존성
# Python 3.10+ 필요
# 설치: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════

# ── 핵심 데이터 처리 ──────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# ── 시각화 ────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# ── 머신러닝 (LSTM 물동량 예측) ───────────────────────────────
scikit-learn>=1.3.0
torch>=2.0.0          # CPU 버전: pip install torch --index-url https://download.pytorch.org/whl/cpu

# ── 웹 데이터 수집 ────────────────────────────────────────────
requests>=2.31.0
feedparser>=6.0.10    # RSS 뉴스 수집 (gCaptain, Splash247, 한국해운신문)

# ── Excel 파일 처리 (KCCI XLS) ────────────────────────────────
xlrd>=2.0.1           # .xls (구형 Excel 97-2003)
openpyxl>=3.1.0       # .xlsx (최신 Excel)

# ── Jupyter 환경 (노트북 발표용) ──────────────────────────────
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
nbformat>=5.9.0
nbconvert>=7.10.0

# ── 테스트 ────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0

# ── Streamlit 웹앱 (app.py) ───────────────────────────────────
streamlit>=1.35.0

# ── FastAPI 백엔드 (api.py — Lovable 연동용) ──────────────────
fastapi>=0.111.0
uvicorn>=0.29.0       # ASGI 서버: uvicorn api:app --reload

# ── LLM 보고서 (둘 중 하나만 설치해도 됨) ─────────────────────
google-generativeai>=0.8.0   # Gemini 1.5 Flash — 무료 (권장)
anthropic>=0.40.0             # Claude Haiku 4.5 — 유료 대안

# ── 지도 시각화 (ODCY 추천 — 없으면 matplotlib fallback) ──────
folium>=0.15.0

# ── 개발 편의 ─────────────────────────────────────────────────
python-dotenv>=1.0.0  # .env API 키 관리
python-docx>=1.1.0    # 문서 자동 생성 (가이드 DOCX)
